reference/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
databricks-sql-connector>=3.0.0
databricks-sdk>=0.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
